+++ zsshlib/ssh.go
@@ -22,6 +22,7 @@ import (
 	"github.com/openziti/sdk-golang/ziti/config"
 	"github.com/pkg/errors"
 	"github.com/pkg/sftp"
+	"io"
 	"io/ioutil"
 	"log"
 	"net"
@@ -200,6 +201,29 @@ func SendFile(client *sftp.Client, localPath string, remotePath string) error {
 	if err != nil {
 		return err
 	}
+
+	return nil
+}
+
+func RetrieveRemoteFiles(client *sftp.Client, localPath string, remotePath string) error {
+
+	rf, err := client.Open(remotePath)
+	if err != nil {
+		return fmt.Errorf("error opening remote file [%s] (%w)", remotePath, err)
+	}
+	defer func() { _ = rf.Close() }()
+
+	lf, err := os.OpenFile(localPath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, os.ModePerm)
+	if err != nil {
+		return fmt.Errorf("error opening local file [%s] (%w)", remotePath, err)
+	}
+	defer func() { _ = lf.Close() }()
+
+	_, err = io.Copy(lf, rf)
+	if err != nil {
+		return fmt.Errorf("error copying remote file to local [%s] (%w)", remotePath, err)
+	}
+	logrus.Infof("%s => %s", remotePath, localPath)
 
 	return nil
 }
